+++ client/src/pages/Cart/styles.module.scss
@@ -21,12 +21,6 @@
   border-radius: 0px 10px 10px 0px;
 }
 
-.button_checkout {
-  border-radius: 10px;
-  padding-inline: 50px;
-  color: white;
-  background-color: $--purple-color;
-}
 .form {
   padding: 20px 30px 37px 0px;
   display: flex;
@@ -56,7 +50,6 @@
   display: grid;
   grid-template-columns: 2fr repeat(5, 1fr);
   padding: 65px 0px 50px 0px;
-
   border-bottom: 1px solid #ddd;
 }
